Basic Inputs Script and Prefab
Committed an Input Script to handle basic player input as well as
prefabed it so that the input handler can go on any scene and can be
called as such.

But that shouldnt be an issue since there will be one Master Game Input
(For In-Game) and another for  Menu. But right now will import
animations and models next for us to play around with