Menu Scripts
Menu Scripts have been updated with full controller support as well as keyboard support.

Pause Menu has been updated and cleaned as well to have full controller support and keyboard support.

Added an archive for scripts that are not used due to integration and combining of scripts to streamline things.